correct velocity display in pV diagrams
APLpy always uses m/s as the velocity unit for position-velocity diagram regardless of what the header is set to. It is very annoying to have 6-7 digit tick labels and there is no way to do this in APLpy. Hacking the plot itself works, though. The only workaround in APLpy is to deliberately make the header wrong so the display is correct.